backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
Pillow==11.0.0
pytesseract==0.3.13
reportlab==4.2.0
boto3==1.35.0
httpx==0.28.0
python-dotenv==1.0.1
